requirements.txt
streamlit
pillow
requests
langchain
groq